% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\name{addChapter,QuartoObject-method}
\alias{addChapter,QuartoObject-method}
\alias{addChapter-QuartoObject}
\title{Add a path to the template search path}
\usage{
\S4method{addChapter}{QuartoObject}(x, file, .after = NA)
}
\description{
Add a path to the template search path
}
