% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\docType{methods}
\name{addChapter}
\alias{addChapter}
\title{Add an Entry from the Chapter list}
\usage{
addChapter(x, file, .after = NA)
}
\arguments{
\item{x}{a QuartoObject object}

\item{file}{the file to be added.  See Usage Notes below.}

\item{.after}{At what position should the new file be added?  The default value, \code{NA}, adds the new file at the end of the existing list}
}
\description{
Add an Entry from the Chapter list
}
\section{Usage Notes}{

The locations that are searched to locate \code{file} at the time the \code{QuartoObject}
is \code{publish}ed depend on the way in which \code{file} is defined.  If \code{file} is a path,
whether relative or absolute, then only that location is searched.  If the file
is specified using only a base name, then the elements of \code{templateSearchPath}
are searched, in order, to locate a file with that base name.  Only the first
match is returned.
}

