% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\name{addTemplatePath,QuartoObject-method}
\alias{addTemplatePath,QuartoObject-method}
\alias{addTemplatePath-QuartoObject}
\title{Add a path to the template search path}
\usage{
\S4method{addTemplatePath}{QuartoObject}(x, path, mustExist = FALSE, .after = NA)
}
\description{
Add a path to the template search path
}
