% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{askForParams}
\alias{askForParams}
\title{Run a Shiny App to request Parameters for a Parameterised Report}
\usage{
askForParams(...)
}
\arguments{
\item{...}{Parameters passed to the app via \code{.GlobalEnv$askForParamsInbput}.
See Usage Notes below.}
}
\description{
When given a Quarto document, website or book, interrogate the input to determine
if it has a YAML header.  If it does, parse the YAML header to identify what,
if any, parameters are required.  Display a GUI to obtain the required information
from the user and then update the Quarto object with the information provided.
}
\section{Usage notes}{

If the function is called from an R session that is not interactive, an error
is thrown. (TODO: Does this affect validation?)
}

\section{Usage Notes}{

Currently, the only parameter supported, and which is required, is \code{path}.
This should define the path to the Quarto document, website or book
}

