% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoquarto-package.R
\docType{package}
\name{autoquarto-package}
\alias{autoquarto}
\alias{autoquarto-package}
\title{autoquarto: Tools to help programmatic compilation of Quarto books}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: John Kirkpatrick \email{john.kirkpatrick@roche.com} (\href{https://orcid.org/0000-0001-7790-3708}{ORCID})

}
\keyword{internal}
