% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-publish.R
\name{.prepareToPublish}
\alias{.prepareToPublish}
\title{Prepare to Publish a Quarto Object}
\usage{
.prepareToPublish(workDir, outFile, logFile, tidyUp)
}
\arguments{
\item{workDir}{The working folder}

\item{outFile}{The name of the output file}

\item{logFile}{The requested log file}

\item{tidyUp}{delete the contents of workDir when done?}
}
\value{
The path to the working directory
}
\description{
Check the working folder is accessible, set up the log file
}
