% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isQuartoBook}
\alias{isQuartoBook}
\title{Determine if a path points to a Quarto book}
\usage{
isQuartoBook(path)
}
\arguments{
\item{path}{the path to be tested}
}
\value{
\code{TRUE} if the path is a valid Quarto book, \code{FALSE} otherwise
}
\description{
The path is a Quarto book if and only if:
\itemize{
\item It is a readable directory
\item It contains a readable file named \verb{_project.yml}
\item \verb{_project.yml} contains a \code{project: type: book} entry
}
}
