% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isQuartoDocument}
\alias{isQuartoDocument}
\title{Determine if a path points to a Quarto document}
\usage{
isQuartoDocument(path)
}
\arguments{
\item{path}{the path to be tested}
}
\value{
\code{TRUE} if the path is a valid Quarto document, \code{FALSE} otherwise
}
\description{
The path is a Quarto document if and only if:
\itemize{
\item It is a readable file with extension "qmd"
}
}
