% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isQuartoObject}
\alias{isQuartoObject}
\title{Determine if a path points to a Quarto object}
\usage{
isQuartoObject(path)
}
\arguments{
\item{path}{the path to be tested}
}
\value{
\code{TRUE} if the path is a valid Quarto object, \code{FALSE} otherwise
}
\description{
Determine if a path points to a Quarto object
}
