% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\name{locateTemplate,QuartoObject-method}
\alias{locateTemplate,QuartoObject-method}
\alias{locateTemplate-QuartoObject}
\title{Search the \code{templateSearchPath} for the First Matching Template File}
\usage{
\S4method{locateTemplate}{QuartoObject}(x, fileName, strict = TRUE, allowedExtensions = c("qmd"))
}
\arguments{
\item{strict}{Boolean.  Default: TRUE.  Check that the file found (a) exists
(b) is readable and (c) has one of the \code{allowedExtensions}}

\item{allowedExtensions}{Default: \code{"qmd"}.  A vector of permitted extensions.
Ignored if \code{strict} is \code{FALSE}}
}
\description{
Search the \code{templateSearchPath} for the First Matching Template File
}
