% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{parameterEditorPanelServer}
\alias{parameterEditorPanelServer}
\title{Provide server functionality required by the Parameter Editor Panel}
\usage{
parameterEditorPanelServer(id, path)
}
\arguments{
\item{id}{the ID parameter passed to \code{shiny::NS()}}

\item{path}{the path of the Quarto book/website/document whose parameters
are to be updated}
}
\description{
Provide server functionality required by the Parameter Editor Panel
}
