% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{parameterEditorPanelUI}
\alias{parameterEditorPanelUI}
\title{Provide a GUI to enable editing of the parameters in a YAML header}
\usage{
parameterEditorPanelUI(id)
}
\arguments{
\item{id}{the ID parameter passed to \code{shiny::NS()}.}
}
\description{
Provide a GUI to enable editing of the parameters in a YAML header
}
