% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymlthis.R
\name{parameterTibbleToYAML}
\alias{parameterTibbleToYAML}
\title{Convert a tibble f parameter definitions to YAML}
\usage{
parameterTibbleToYAML(d)
}
\arguments{
\item{d}{The parameter tibble}
}
\value{
a YML object
}
\description{
Convert a tibble f parameter definitions to YAML
}
\section{Usage Notes}{

The input tibble should contain columns Parameter, Value, Mode, IsNULL and IsNA.
}

