% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\name{publish,QuartoCompoundObject-method}
\alias{publish,QuartoCompoundObject-method}
\alias{publish-QuartoCompoundObject}
\title{Publish a QuartoCompoundObject object}
\usage{
\S4method{publish}{QuartoCompoundObject}(
  x,
  outFile,
  params = list(),
  workDir = NULL,
  tidyUp = TRUE,
  logFile = NA
)
}
\arguments{
\item{params}{Default \code{list()}.  A list containing the parameters to use when
publishing the QuartoCompoundObject}

\item{workDir}{Default: \code{NULL}.  The temporary work folder used to generate the report.
If \code{NULL} a temporary work folder with a random name, is used.}

\item{tidyUp}{Boolean.  Should the workDir be deleted on successful publication?}

\item{logFile}{Default: NA.  The name of the log file that documents the publishing process.
If \code{NA}, the name is based on \code{basename(outFile)}, with the start date/time appended and
located in the folder returned by \code{here::here()} at the time the function was called.
If \code{NULL}, no log file is produced.}
}
\description{
Publish a QuartoCompoundObject object
}
\examples{
\dontrun{
publish(
  QuartoBook(
    templateSearchPath = list(testthat::test_path("testData", "global")),
    chapters = list(
      "index.qmd",
      testthat::test_path("testData", "testParams", "introParams.qmd")
    )
  ),
  outFile = "testParamsOutput",
  workDir = testthat::test_path("testData", "_work"),
  params = params
)
}
}
