% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\name{publish,QuartoDocument-method}
\alias{publish,QuartoDocument-method}
\alias{publish-QuartoDocument}
\title{Publish a QuartoDocument object}
\usage{
\S4method{publish}{QuartoDocument}(
  x,
  outFile,
  params = list(),
  workDir = NULL,
  tidyUp = FALSE,
  logFile = NA,
  ...
)
}
\arguments{
\item{params}{Default \code{list()}.  A list containing the parameters to use when
publishing the QuartoDocument}

\item{workDir}{Default: \code{NULL}.  The temporary work folder used to generate the report.
If \code{NULL} a temporary work folder with a random name, is used.}

\item{tidyUp}{Boolean.  Should the workDir be deleted on successful publication?}

\item{logFile}{Default: NA.  The name of the log file that documents the publishing process.
If \code{NA}, the name is based on \code{basename(outFile)}, with the start date/time appended and
located in the folder returned by \code{here::here()} at the time the function was called.
If \code{NULL}, no log file is produced.}

\item{...}{passed to \code{quarto::quarto_render}}
}
\description{
Publish a QuartoDocument object
}
\examples{
\dontrun{
publish(
  QuartoDocument(testthat::test_path("testData", "testParams", "introParams.qmd")),
  outFile = "testParamsOutputDocument",
  workDir = testthat::test_path("testData", "_work"),
  params = list(
    dataName = "mtcars",
    rowCount = 10,
    x = "wt",
    y = "mpg",
    g = "cyl",
    plotTitle = "A title for the plot"
  )
)
}
}
