% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartobook-class.R
\name{QuartoBook}
\alias{QuartoBook}
\title{Constructor for the \code{QuartoBook} class}
\usage{
QuartoBook(
  templateSearchPath = list("."),
  variables = list(),
  chapters = list()
)
}
\arguments{
\item{templateSearchPath}{the initial value of the \code{templateSearchPath} slot}

\item{variables}{the initial value of the \code{variables} slot}

\item{chapters}{the initial value of the \code{chapters} slot}
}
\description{
Constructor for the \code{QuartoBook} class
}
