% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-class.R
\docType{class}
\name{QuartoCompoundObject-class}
\alias{QuartoCompoundObject-class}
\title{Base Class for Compound Quarto Objects}
\description{
Compound Quarto objects are books and websites: they contain chapters
}
\section{Slots}{

\describe{
\item{\code{templateSearchPath}}{a \code{list} containing the paths to be searched for the
template files that define the \code{chapters} of this object.}

\item{\code{variables}}{a list containing the names (and values) of the variables required to
create the book}

\item{\code{chapters}}{a list containing the names of the quarto documents that comprise the book}

\item{\code{type}}{the type of the object (currently book or website, depending on subclass)}
}}

