% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-class.R
\name{QuartoDocument}
\alias{QuartoDocument}
\title{Constructor for the \code{QuartoDocument} class}
\usage{
QuartoDocument(fileName, variables = list())
}
\arguments{
\item{fileName}{the Quarto file represented by this object}

\item{variables}{the initial value of the \code{variables} slot}
}
\description{
Constructor for the \code{QuartoDocument} class
}
