% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\name{quartoYML,QuartoObject-method}
\alias{quartoYML,QuartoObject-method}
\alias{quartoYML-QuartoObject}
\title{Obtain the _quarto.yml Component of the Quarto Object}
\usage{
\S4method{quartoYML}{QuartoObject}(x)
}
\description{
Obtain the _quarto.yml Component of the Quarto Object
}
