% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\docType{methods}
\name{quartoYML}
\alias{quartoYML}
\title{Obtain the _quarto.yml Component of the Quarto Object}
\usage{
quartoYML(x, ...)
}
\arguments{
\item{x}{a QuartoObject object}

\item{...}{passed to class methods}
}
\description{
Obtain the _quarto.yml Component of the Quarto Object
}
