% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymlthis.R
\name{readYamlFrontMatter}
\alias{readYamlFrontMatter}
\title{Read the YAML Front Matter From a File}
\usage{
readYamlFrontMatter(path, item = NA)
}
\arguments{
\item{path}{The path to the file whose front matter is to be read}

\item{item}{If not NA, the element of the front matter to be returned}
}
\value{
a list containing the YAML front matter
}
\description{
\code{rmarkdown::yaml_front_matter} does not work correctly when the parameter name
is \code{y}, \code{Y}, \code{Yes}, \code{n} and possibly other similar values.  \code{ymlthis::yml_pluck}
errors when the parameter value is an unquoted integer.  Both problems seem to
be related to the fact that \code{yaml::read_yaml} uses version 1.1 of the YAML spec
rather than version 1.2.
}
\details{
So use this internal method as a way to avoid both issues.
}
