% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\name{removeChapter,QuartoWebsite-method}
\alias{removeChapter,QuartoWebsite-method}
\alias{removeChapter-QuartoWebsite}
\title{Remove a chapter from the chapter list}
\usage{
\S4method{removeChapter}{QuartoWebsite}(x, file = NA, pos = NA)
}
\description{
Remove a chapter from the chapter list
}
