% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\name{removeTemplatePath,QuartoObject-method}
\alias{removeTemplatePath,QuartoObject-method}
\alias{removeTemplatePath-QuartoWebsite}
\title{Remove a path from the template search path}
\usage{
\S4method{removeTemplatePath}{QuartoObject}(x, path = NA, pos = NA)
}
\description{
Remove a path from the template search path
}
