% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\docType{methods}
\name{removeTemplatePath}
\alias{removeTemplatePath}
\alias{removeTemplatePath,QuartoObject-method}
\alias{removeTemplatePath-QuartoWebsite}
\title{Remove an Entry from the Template Search Path list}
\usage{
removeTemplatePath(x, path = NA, pos = NA)

\S4method{removeTemplatePath}{QuartoObject}(x, path = NA, pos = NA)
}
\arguments{
\item{x}{a QuartoObject object}

\item{path}{the path to be removed}

\item{pos}{the position of the path to be removed}
}
\description{
Remove an Entry from the Template Search Path list

Remove a path from the template search path
}
\section{Methods (by class)}{
\itemize{
\item \code{removeTemplatePath(QuartoObject)}: 

}}
\section{Usage Notes}{

Either \code{path} or \code{pos} must be provided, but not both
}

