% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymlthis.R
\name{replaceYamlParams}
\alias{replaceYamlParams}
\title{Replace the params Element of a File's YAML Front Matter}
\usage{
replaceYamlParams(path, params, restrict = c("none", "both", "file", "list"))
}
\arguments{
\item{path}{The path to the file whose front matter is to be updated}

\item{params}{The new parameter values}

\item{restrict}{How to restrict the updating of the parameters.  \code{"none"}
results in any parameter that is in either the front matter or the list to be
written to the edited file.  \code{"both"} means only those parameters that are in
both the original front matter and the list are included.  \code{"file"} means that
only parameter names in the original front matter are included and those only
in the list are ignored.  \code{"list"} means that only those parameters in the
list are included and those only in the original front matter are ignored.}
}
\description{
Replace the params Element of a File's YAML Front Matter
}
