% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymlthis.R
\name{yml_book}
\alias{yml_book}
\title{Create a YAML book element}
\usage{
yml_book(.yml, ...)
}
\arguments{
\item{.yml}{a \code{yml} object created by \code{yml()}, \code{as_yml()}, or returned by
a \verb{yml_*()} function}

\item{...}{valid R code calling functions that return objects of class
\code{rmarkdown_output_format}, such as the \verb{*_document()} functions in
rmarkdown.}
}
\value{
a \code{yml} object
}
\description{
Create a YAML book element
}
