% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymlthis.R
\name{yml_element}
\alias{yml_element}
\title{Capture, validate, and write YAML}
\usage{
yml_element(.yml, name, ...)
}
\arguments{
\item{.yml}{a \code{yml} object created by \code{yml()}, \code{as_yml()}, or returned by
a \verb{yml_*()} function}

\item{name}{the name of the YAML field being created}

\item{...}{valid R code calling functions that return objects of class
\code{rmarkdown_output_format}, such as the \verb{*_document()} functions in
rmarkdown.}
}
\value{
a \code{yml} object
}
\description{
\code{yml_element()} writes valid YAML for an arbitrary field of R Markdown YAML.
\code{yml_element()} captures the actual field functions and translates them to YAML.
This function accepts multiple output formats (separated by commas) and
validates each by evaluating the function internally.
The YAML fields under the parent element come from arguments in their
respective R functions. If you wanted to see the available fields in
\code{pdf_document()}, for instance, you would read the documentation for that
function using \code{?pdf_document}.
}
\examples{
identical(
  ymlthis::yml_empty() \%>\% yml_element("output", html_document()), 
  ymlthis::yml_empty() \%>\% ymlthis::yml_output(html_document()) # TRUE
)
}
