% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-publish.R
\name{.processProjectYAML}
\alias{.processProjectYAML}
\title{Process a project YAML}
\usage{
.processProjectYAML(x, workDir, outFile, params = list())
}
\arguments{
\item{x}{The QuartoCompoundObject that defines the project YAML}

\item{workDir}{The working directory}

\item{outFile}{The path to the output file}

\item{params}{A list of parameters to write to the project YAML}
}
\value{
the text written to \verb{_quarto.yml}, invisibly
}
\description{
Create the _quarto.yml for a QuartoCompoundObject
}
