% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-publish.R
\name{.processProjectYAML}
\alias{.processProjectYAML}
\title{Process a project YAML}
\usage{
.processProjectYAML(x, workDir, outFile)
}
\arguments{
\item{x}{The QuartoObject that defines the project YAML}

\item{workDir}{The working directory}

\item{outFile}{The path to the ouytput file}
}
\description{
Process a project YAML
}
