% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-publish.R
\name{.renderQuartoCompoundObject}
\alias{.renderQuartoCompoundObject}
\title{Render a CompoundQuartoObject}
\usage{
.renderQuartoCompoundObject(workDir)
}
\arguments{
\item{The}{working directory containing the _quarto.yaml file}
}
\description{
Render a CompoundQuartoObject
}
