% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\docType{methods}
\name{locateTemplate}
\alias{locateTemplate}
\alias{locateTemplate,QuartoObject-method}
\alias{locateTemplate-QuartoObject}
\title{Search the \code{templateSearchPath} for the First Matching Template File}
\usage{
locateTemplate(x, fileName = NA, ...)

\S4method{locateTemplate}{QuartoObject}(x, fileName, strict = TRUE, allowedExtensions = c("qmd"))
}
\arguments{
\item{x}{a QuartoObject object}

\item{fileName}{the (base) name of the file to be searched for along the
\code{templateSearchPath} defined by \code{x}.}

\item{...}{passed to class methods}

\item{strict}{Boolean.  Default: TRUE.  Check that the file found (a) exists
(b) is readable and (c) has one of the \code{allowedExtensions}}

\item{allowedExtensions}{Default: \code{"qmd"}.  A vector of permitted extensions.
Ignored if \code{strict} is \code{FALSE}}
}
\description{
Search the \code{templateSearchPath} for the First Matching Template File

Search the \code{templateSearchPath} for the First Matching Template File
}
\section{Usage Notes}{

If \code{basename(fileName) == fileName} then \code{x@templateSearchPath} is searched, in order, for
a file with the given name.  The first match is returned.  Otherwise,
the absolute path to the given file is returned.  This process allows (say) global,
project and study level templates to be obtained.
}

