% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\docType{methods}
\name{publish}
\alias{publish}
\title{Publish a Quarto object}
\usage{
publish(x, outFile, ...)
}
\arguments{
\item{x}{a QuartoObject object}

\item{outFile}{the name of the file containing the published Quarto object}

\item{...}{passed to class methods}
}
\description{
Publish a Quarto object
}
