% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\docType{methods}
\name{publish}
\alias{publish}
\alias{publish,QuartoObject-method}
\alias{publish-QuartoObject}
\alias{quartoYML,QuartoObject-method}
\alias{quartoYML-QuartoObject}
\title{Publish a Quarto object}
\usage{
publish(x, outFile, ...)

\S4method{publish}{QuartoObject}(
  x,
  outFile,
  params = list(),
  workDir = NULL,
  tidyUp = FALSE,
  logFile = NA
)

\S4method{quartoYML}{QuartoObject}(x)
}
\arguments{
\item{x}{a QuartoObject object}

\item{outFile}{the name of the file containing the published Quarto object}

\item{...}{passed to class methods}

\item{params}{Default \code{list()}.  A list containing the parameters to use when
publishing the QuartoObject}

\item{workDir}{Default: \code{NULL}.  The temporary work folder used to generate the report.
If \code{NULL} a temporary work folder with a random name, is used.}

\item{tidyUp}{Boolean.  Should the workDir be deleted on successful publication?}

\item{logFile}{Default: NA.  The name of the log file that documents the publishing process.
If \code{NA}, the name is based on \code{basename(outFile)}, with the start date/time appended and
located in the folder returned by \code{here::here()} at the time the function was called.
If \code{NULL}, no log file is produced.}
}
\description{
Publish a Quarto object

Publish a QuartoObject object

Obtain the _quarto.yml Component of the Quarto Object
}
\section{Methods (by class)}{
\itemize{
\item \code{publish(QuartoObject)}: 

\item \code{quartoYML(QuartoObject)}: 

}}
