% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-class.R
\docType{class}
\name{QuartoObject-class}
\alias{QuartoObject-class}
\title{Base Class for All Quarto Objects}
\description{
Base Class for All Quarto Objects
}
\section{Slots}{

\describe{
\item{\code{variables}}{a list containing the names (and values) of the variables required to
create the book}
}}

