% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\docType{methods}
\name{removeChapter}
\alias{removeChapter}
\alias{removeChapter,QuartoWebsite-method}
\alias{removeChapter-QuartoWebsite}
\title{Remove an Entry from the Chapter list}
\usage{
removeChapter(x, file = NA, pos = NA)

\S4method{removeChapter}{QuartoWebsite}(x, file = NA, pos = NA)
}
\arguments{
\item{x}{a QuartoObject object}

\item{file}{the file to be removed}

\item{pos}{the position of the file to be removed}
}
\description{
Remove an Entry from the Chapter list

Remove a chapter from the chapter list
}
\section{Usage Notes}{

Either \code{path} or \code{pos} must be provided, but not both
}

