% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartoobject-methods.R
\docType{methods}
\name{removeTemplatePath}
\alias{removeTemplatePath}

\alias{removeTemplatePath,QuartoCompoundObject-method}
\alias{removeTemplatePath-QuartoCompoundObject}
\title{Remove an Entry from the Template Search Path list}
\usage{
removeTemplatePath(x, path = NA, pos = NA)

\S4method{removeTemplatePath}{QuartoCompoundObject}(x, path = NA, pos = NA)
}
\arguments{
\item{x}{a QuartoCompoundObject object}

\item{path}{the path to be removed}

\item{pos}{the position of the path to be removed}
}
\description{
Remove an Entry from the Template Search Path list
}
\section{Usage Notes}{

Either \code{path} or \code{pos} must be provided, but not both
}

