% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymlthis.R
\name{yml_project}
\alias{yml_project}
\title{Create a YAML project Element}
\usage{
yml_project(.project, ..., validNames = c("part"))
}
\value{
a \code{yml} object
}
\description{
Create a YAML project Element
}
